+++ spec/mailers/user_mailer_spec.rb
@@ -2,6 +2,10 @@ require 'rails_helper'
 
 RSpec.describe UserMailer, type: :mailer do
   # アカウント削除受け付けのお知らせ
+  # 前提条件
+  #   削除予約済み  
+  # テストパターン
+  #   なし
   describe 'destroy_reserved' do
     let(:user) { FactoryBot.create(:user, destroy_schedule_at: Time.current + Settings['destroy_schedule_days'].days) }
     let(:mail) { UserMailer.with(user: user).destroy_reserved }
@@ -12,14 +16,18 @@ RSpec.describe UserMailer, type: :mailer do
       expect(mail.to).to eq([user.email])
     end
     it '本文(html)にアカウント削除取り消しのURLが含まれる' do
-      expect(mail.html_part.body).to include("\"#{users_undo_delete_url}\"")
+      expect(mail.html_part.body).to include("\"#{delete_undo_user_registration_url}\"")
     end
     it '本文(text)にアカウント削除取り消しのURLが含まれる' do
-      expect(mail.text_part.body).to include(users_undo_delete_url)
+      expect(mail.text_part.body).to include(delete_undo_user_registration_url)
     end
   end
 
   # アカウント削除取り消し完了のお知らせ
+  # 前提条件
+  #   なし
+  # テストパターン
+  #   なし
   describe 'undo_destroy_reserved' do
     let(:user) { FactoryBot.create(:user) }
     let(:mail) { UserMailer.with(user: user).undo_destroy_reserved }
@@ -32,6 +40,10 @@ RSpec.describe UserMailer, type: :mailer do
   end
 
   # アカウント削除完了のお知らせ
+  # 前提条件
+  #   なし
+  # テストパターン
+  #   なし
   describe 'destroy_completed' do
     let(:user) { FactoryBot.create(:user) }
     let(:mail) { UserMailer.with(user: user).destroy_completed }