+++ spec/requests/users/passwords_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe 'Users::Passwords', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'GET /new' do
+  describe 'GET #new' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
@@ -37,27 +37,27 @@ RSpec.describe 'Users::Passwords', type: :request do
     end
   end
 
-  # POST /users/password パスワード再設定[メール送信](処理)
+  # POST /users/password/new パスワード再設定[メール送信](処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   有効なパラメータ, 無効なパラメータ → 事前にデータ作成
-  describe 'POST /create' do
+  describe 'POST #create' do
     let!(:send_user) { FactoryBot.create(:user) }
     let!(:valid_attributes) { FactoryBot.attributes_for(:user, email: send_user.email) }
     let!(:invalid_attributes) { FactoryBot.attributes_for(:user, email: nil) }
 
     # テスト内容
-    shared_examples_for 'ToOK' do
-      it '成功ステータス' do
-        post user_password_path, params: { user: attributes }
+    shared_examples_for 'ToError' do
+      it '成功ステータス' do # Tips: 再入力
+        post create_user_password_path, params: { user: attributes }
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        post user_password_path, params: { user: attributes }
+        post create_user_password_path, params: { user: attributes }
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -65,7 +65,7 @@ RSpec.describe 'Users::Passwords', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        post user_password_path, params: { user: attributes }
+        post create_user_password_path, params: { user: attributes }
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -83,7 +83,7 @@ RSpec.describe 'Users::Passwords', type: :request do
     end
     shared_examples_for '[未ログイン]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
-      it_behaves_like 'ToOK' # Tips: 再入力
+      it_behaves_like 'ToError'
     end
     shared_examples_for '[ログイン中/削除予約済み]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
@@ -112,7 +112,7 @@ RSpec.describe 'Users::Passwords', type: :request do
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   トークン: 期限内, 期限切れ, 存在しない, ない → データ作成
-  describe 'GET /edit' do
+  describe 'GET #edit' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
@@ -201,40 +201,40 @@ RSpec.describe 'Users::Passwords', type: :request do
     end
   end
 
-  # PUT /users/password パスワード再設定(処理)
+  # PUT(PATCH) /users/password パスワード再設定(処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   トークン: 期限内, 期限切れ, 存在しない, ない → データ作成
   #   有効なパラメータ, 無効なパラメータ → 事前にデータ作成
-  describe 'PUT /update' do
+  describe 'PUT #update' do
     let!(:valid_attributes) { FactoryBot.attributes_for(:user) }
     let!(:invalid_attributes) { FactoryBot.attributes_for(:user, password: nil) }
 
     # テスト内容
     shared_examples_for 'OK' do
       it 'パスワードリセット送信日時がなしに変更される' do
-        put user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
+        put update_user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
         expect(User.find(@send_user.id).reset_password_sent_at).to be_nil
       end
     end
     shared_examples_for 'NG' do
       it 'パスワードリセット送信日時が変更されない' do
-        put user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
+        put update_user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
         expect(User.find(@send_user.id).reset_password_sent_at).to eq(@send_user.reset_password_sent_at)
       end
     end
 
-    shared_examples_for 'ToOK' do
-      it '成功ステータス' do
-        put user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
+    shared_examples_for 'ToError' do
+      it '成功ステータス' do # Tips: 再入力
+        put update_user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        put user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
+        put update_user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -242,7 +242,7 @@ RSpec.describe 'Users::Passwords', type: :request do
     end
     shared_examples_for 'ToNew' do |alert, notice|
       it 'パスワード再設定[メール送信]にリダイレクト' do
-        put user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
+        put update_user_password_path, params: { user: attributes.merge({ reset_password_token: reset_password_token }) }
         expect(response).to redirect_to(new_user_password_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -278,7 +278,7 @@ RSpec.describe 'Users::Passwords', type: :request do
     shared_examples_for '[未ログイン][期限内]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
       it_behaves_like 'NG'
-      it_behaves_like 'ToOK' # Tips: 再入力
+      it_behaves_like 'ToError'
     end
     shared_examples_for '[ログイン中/削除予約済み][期限内/期限切れ]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }