+++ spec/requests/users/registrations_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'GET /new' do
+  describe 'GET #new' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
@@ -37,13 +37,13 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
   end
 
-  # POST /users アカウント登録(処理)
+  # POST /users/sign_up アカウント登録(処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   有効なパラメータ, 無効なパラメータ → 事前にデータ作成
-  describe 'POST /create' do
+  describe 'POST #create' do
     let!(:valid_attributes) { FactoryBot.attributes_for(:user) }
     let!(:invalid_attributes) { FactoryBot.attributes_for(:user, email: nil) }
 
@@ -51,27 +51,27 @@ RSpec.describe 'Users::Registrations', type: :request do
     shared_examples_for 'OK' do
       it '作成される' do
         expect do
-          post user_registration_path, params: { user: attributes }
+          post create_user_registration_path, params: { user: attributes }
         end.to change(User, :count).by(1)
       end
     end
     shared_examples_for 'NG' do
       it '作成されない' do
         expect do
-          post user_registration_path, params: { user: attributes }
+          post create_user_registration_path, params: { user: attributes }
         end.to change(User, :count).by(0)
       end
     end
 
-    shared_examples_for 'ToOK' do
-      it '成功ステータス' do
-        post user_registration_path, params: { user: attributes }
+    shared_examples_for 'ToError' do
+      it '成功ステータス' do # Tips: 再入力
+        post create_user_registration_path, params: { user: attributes }
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        post user_registration_path, params: { user: attributes }
+        post create_user_registration_path, params: { user: attributes }
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -79,7 +79,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        post user_registration_path, params: { user: attributes }
+        post create_user_registration_path, params: { user: attributes }
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -100,7 +100,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     shared_examples_for '[未ログイン]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
       it_behaves_like 'NG'
-      it_behaves_like 'ToOK' # Tips: 再入力
+      it_behaves_like 'ToError'
     end
     shared_examples_for '[ログイン中/削除予約済み]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
@@ -129,7 +129,7 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'GET /edit' do
+  describe 'GET #edit' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
@@ -168,39 +168,39 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
   end
 
-  # PUT /users 登録情報変更(処理)
+  # PUT(PATCH) /users/edit 登録情報変更(処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   有効なパラメータ, 無効なパラメータ → 事前にデータ作成
-  describe 'PUT /update' do
+  describe 'PUT #update' do
     let!(:valid_attributes) { FactoryBot.attributes_for(:user) }
     let!(:invalid_attributes) { FactoryBot.attributes_for(:user, email: nil) }
 
     # テスト内容
     shared_examples_for 'OK' do
       it '氏名が変更される' do
-        put user_registration_path, params: { user: attributes.merge(current_password: current_password) }
+        put update_user_registration_path, params: { user: attributes.merge(current_password: current_password) }
         expect(User.find(user.id).name).to eq(attributes[:name])
       end
     end
     shared_examples_for 'NG' do
       it '氏名が変更されない' do
-        put user_registration_path, params: { user: attributes.merge(current_password: current_password) }
+        put update_user_registration_path, params: { user: attributes.merge(current_password: current_password) }
         expect(User.find(user.id).name).to eq(user.name)
       end
     end
 
-    shared_examples_for 'ToOK' do
-      it '成功ステータス' do
-        put user_registration_path, params: { user: attributes.merge(current_password: current_password) }
+    shared_examples_for 'ToError' do
+      it '成功ステータス' do # Tips: 再入力
+        put update_user_registration_path, params: { user: attributes.merge(current_password: current_password) }
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        put user_registration_path, params: { user: attributes.merge(current_password: current_password) }
+        put update_user_registration_path, params: { user: attributes.merge(current_password: current_password) }
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -208,7 +208,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        put user_registration_path, params: { user: attributes.merge(current_password: current_password) }
+        put update_user_registration_path, params: { user: attributes.merge(current_password: current_password) }
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -239,7 +239,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     shared_examples_for '[ログイン中]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
       it_behaves_like 'NG'
-      it_behaves_like 'ToOK' # Tips: 再入力
+      it_behaves_like 'ToError'
     end
     shared_examples_for '[削除予約済み]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
@@ -270,20 +270,20 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
   end
 
-  # PUT /users/image 画像変更(処理)
+  # PUT(PATCH) /users/image 画像変更(処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   有効なパラメータ, 無効なパラメータ → 事前にデータ作成
-  describe 'PUT /image_update' do
+  describe 'PUT #image_update' do
     let!(:valid_attributes) { { image: fixture_file_upload(TEST_IMAGE_FILE, TEST_IMAGE_TYPE) } }
     let!(:invalid_attributes) { nil }
 
     # テスト内容
     shared_examples_for 'OK' do
       it '画像が変更される' do
-        put users_image_path, params: { user: attributes }
+        put update_user_image_registration_path, params: { user: attributes }
         after_user = User.find(user.id)
         expect(after_user.image.url).not_to eq(user.image.url)
         after_user.remove_image!
@@ -292,20 +292,20 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'NG' do
       it '画像が変更されない' do
-        put users_image_path, params: { user: attributes }
+        put update_user_image_registration_path, params: { user: attributes }
         expect(User.find(user.id).image.url).to eq(user.image.url)
       end
     end
 
-    shared_examples_for 'ToOK' do
-      it '成功ステータス' do
-        put users_image_path, params: { user: attributes }
+    shared_examples_for 'ToError' do
+      it '成功ステータス' do # Tips: 再入力
+        put update_user_image_registration_path, params: { user: attributes }
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        put users_image_path, params: { user: attributes }
+        put update_user_image_registration_path, params: { user: attributes }
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -313,7 +313,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        put users_image_path, params: { user: attributes }
+        put update_user_image_registration_path, params: { user: attributes }
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -321,7 +321,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToEdit' do |alert, notice|
       it '登録情報変更にリダイレクト' do
-        put users_image_path, params: { user: attributes }
+        put update_user_image_registration_path, params: { user: attributes }
         expect(response).to redirect_to(edit_user_registration_path)
         after_user = User.find(user.id)
         after_user.remove_image!
@@ -355,7 +355,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     shared_examples_for '[ログイン中]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
       it_behaves_like 'NG'
-      it_behaves_like 'ToOK' # Tips: 再入力
+      it_behaves_like 'ToError'
     end
     shared_examples_for '[削除予約済み]無効なパラメータ' do
       let!(:attributes) { invalid_attributes }
@@ -384,24 +384,24 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'DELETE /image_destroy' do
+  describe 'DELETE #image_destroy' do
     # テスト内容
     shared_examples_for 'OK' do
       it '画像が削除される' do
-        delete users_image_path
+        delete delete_user_image_registration_path
         expect(User.find(user.id).image.url).to be_nil
       end
     end
     shared_examples_for 'NG' do
       it '画像が変更されない' do
-        delete users_image_path
+        delete delete_user_image_registration_path
         expect(User.find(user.id).image.url).to eq(user.image.url)
       end
     end
 
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        delete users_image_path
+        delete delete_user_image_registration_path
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -409,7 +409,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        delete users_image_path
+        delete delete_user_image_registration_path
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -417,7 +417,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToEdit' do |alert, notice|
       it '登録情報変更にリダイレクト' do
-        delete users_image_path
+        delete delete_user_image_registration_path
         expect(response).to redirect_to(edit_user_registration_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -450,17 +450,17 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'GET /delete' do
+  describe 'GET #delete' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
-        get users_delete_path
+        get delete_user_registration_path
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        get users_delete_path
+        get delete_user_registration_path
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -468,7 +468,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        get users_delete_path
+        get delete_user_registration_path
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -489,21 +489,21 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
   end
 
-  # DELETE /users アカウント削除(処理)
+  # DELETE /users/delete アカウント削除(処理)
   # 前提条件
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'DELETE /destroy' do
+  describe 'DELETE #destroy' do
     # テスト内容
     shared_examples_for 'OK' do
       let!(:start_time) { Time.current - 1.second }
       it '削除依頼日時が現在日時に変更される' do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(user.destroy_requested_at).to be_between(start_time, Time.current)
       end
       it "削除予定日時が#{Settings['destroy_schedule_days']}日後に変更される" do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(user.destroy_schedule_at).to be_between(start_time + Settings['destroy_schedule_days'].days,
                                                        Time.current + Settings['destroy_schedule_days'].days)
       end
@@ -511,18 +511,18 @@ RSpec.describe 'Users::Registrations', type: :request do
     shared_examples_for 'NG' do
       let!(:before_user) { user }
       it '削除依頼日時が変更されない' do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(user.destroy_requested_at).to eq(before_user.destroy_requested_at)
       end
       it '削除予定日時が変更されない' do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(user.destroy_schedule_at).to eq(before_user.destroy_schedule_at)
       end
     end
 
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -530,7 +530,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        delete user_registration_path
+        delete destroy_user_registration_path
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -559,17 +559,17 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'GET /undo_delete' do
+  describe 'GET #undo_delete' do
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do
-        get users_undo_delete_path
+        get delete_undo_user_registration_path
         expect(response).to be_successful
       end
     end
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        get users_undo_delete_path
+        get delete_undo_user_registration_path
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -577,7 +577,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        get users_undo_delete_path
+        get delete_undo_user_registration_path
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -603,33 +603,33 @@ RSpec.describe 'Users::Registrations', type: :request do
   #   なし
   # テストパターン
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
-  describe 'DELETE /undo_destroy' do
+  describe 'DELETE #undo_destroy' do
     # テスト内容
     shared_examples_for 'OK' do
       it '削除依頼日時がなしに変更される' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(user.destroy_requested_at).to be_nil
       end
       it '削除予定日時がなしに変更される' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(user.destroy_schedule_at).to be_nil
       end
     end
     shared_examples_for 'NG' do
       let!(:before_user) { user }
       it '削除依頼日時が変更されない' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(user.destroy_requested_at).to eq(before_user.destroy_requested_at)
       end
       it '削除予定日時が変更されない' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(user.destroy_schedule_at).to eq(before_user.destroy_schedule_at)
       end
     end
 
     shared_examples_for 'ToTop' do |alert, notice|
       it 'トップページにリダイレクト' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(response).to redirect_to(root_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil
@@ -637,7 +637,7 @@ RSpec.describe 'Users::Registrations', type: :request do
     end
     shared_examples_for 'ToLogin' do |alert, notice|
       it 'ログインにリダイレクト' do
-        delete users_undo_delete_path
+        delete destroy_undo_user_registration_path
         expect(response).to redirect_to(new_user_session_path)
         expect(flash[:alert]).to alert.present? ? eq(I18n.t(alert)) : be_nil
         expect(flash[:notice]).to notice.present? ? eq(I18n.t(notice)) : be_nil