+++ spec/routing/admin_users/omniauth_callbacks_routing_spec.rb
@@ -7,12 +7,16 @@ RSpec.describe AdminUsers::OmniauthCallbacksController, type: :routing do
       # expect(post: '/admin_users/auth/twitter').to route_to('admin_users/omniauth_callbacks#passthru')
       expect(get: '/admin_users/auth/twitter').not_to be_routable
       expect(post: '/admin_users/auth/twitter').not_to be_routable
+      # expect(get: '/admin/auth/twitter').not_to be_routable # Tips: rails_admin/main#show
+      expect(post: '/admin/auth/twitter').not_to be_routable
     end
     it 'routes to #failure' do
       # expect(get: '/admin_users/auth/twitter/callback').to route_to('admin_users/omniauth_callbacks#failure')
       # expect(post: '/admin_users/auth/twitter/callback').to route_to('admin_users/omniauth_callbacks#failure')
       expect(get: '/admin_users/auth/twitter/callback').not_to be_routable
       expect(post: '/admin_users/auth/twitter/callback').not_to be_routable
+      expect(get: '/admin/auth/twitter/callback').not_to be_routable
+      expect(post: '/admin/auth/twitter/callback').not_to be_routable
     end
   end
 end